backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9
duckdb>=0.10.0
pandas>=2.2.0
pyarrow>=11.0.0
openpyxl>=3.1.5
fastavro>=1.12.1
